refactored code of the reconstruction plan
