+++ src/components/program/programPlanAdminModals/formItem/ProductLevelItem.tsx
@@ -1,0 +1,42 @@
+import { Spinner } from '@chakra-ui/react'
+import { Form, InputNumber } from 'antd'
+import { useApp } from 'lodestar-app-element/src/contexts/AppContext'
+import { useEffect } from 'react'
+import { useIntl } from 'react-intl'
+import { useProductLevel } from '../../../../hooks/data'
+import programMessages from '../../translation'
+
+interface ParticipantsProps {
+  label?: string
+  name: string
+  programPlanId: string | undefined
+  setProductLevel: React.Dispatch<React.SetStateAction<number>>
+}
+
+const ParticipantsItem: React.FC<ParticipantsProps> = ({ label, name, programPlanId, setProductLevel }) => {
+  const { enabledModules } = useApp()
+  const { loading: loadingProductLevel, productLevel } = useProductLevel(`ProgramPlan_${programPlanId}`)
+
+  const { formatMessage } = useIntl()
+  const _label = label ? label : formatMessage(programMessages.ProgramPlanAdminModal.productLevel)
+
+  useEffect(() => {
+    !loadingProductLevel && setProductLevel(productLevel)
+  }, [setProductLevel, loadingProductLevel, productLevel])
+
+  return (
+    <>
+      {enabledModules.product_level ? (
+        loadingProductLevel ? (
+          <Spinner />
+        ) : (
+          <Form.Item label={_label} name={name}>
+            <InputNumber />
+          </Form.Item>
+        )
+      ) : null}
+    </>
+  )
+}
+
+export default ParticipantsItem
